Restructure core/base slightly
Extra all series related functions into series.go. Keep all offenders in
1 place!

Remove some unused functions and make some functions private

There are now only two (public) functions which parse series,
GetSeriesFromBase and GetBaseFromSeries
