+++ core/base/base.go
@@ -131,46 +131,3 @@ func (b Base) DisplayString() string {
 	}
 	return b.OS + "@" + b.Channel.DisplayString()
 }
-
-// GetBaseFromSeries returns the Base infor for a series.
-func GetBaseFromSeries(series string) (Base, error) {
-	var result Base
-	osName, err := GetOSFromSeries(series)
-	if err != nil {
-		return result, errors.NotValidf("series %q", series)
-	}
-	osVersion, err := SeriesVersion(series)
-	if err != nil {
-		return result, errors.NotValidf("series %q", series)
-	}
-	result.OS = strings.ToLower(osName.String())
-	result.Channel = MakeDefaultChannel(osVersion)
-	return result, nil
-}
-
-// GetSeriesFromChannel gets the series from os name and channel.
-func GetSeriesFromChannel(name string, channel string) (string, error) {
-	base, err := ParseBase(name, channel)
-	if err != nil {
-		return "", errors.Trace(err)
-	}
-	return GetSeriesFromBase(base)
-}
-
-// GetSeriesFromBase returns the series name for a
-// given Base. This is needed to support legacy series.
-func GetSeriesFromBase(v Base) (string, error) {
-	var osSeries map[SeriesName]seriesVersion
-	switch strings.ToLower(v.OS) {
-	case UbuntuOS:
-		osSeries = ubuntuSeries
-	case CentosOS:
-		osSeries = centosSeries
-	}
-	for s, vers := range osSeries {
-		if vers.Version == v.Channel.Track {
-			return string(s), nil
-		}
-	}
-	return "", errors.NotFoundf("os %q version %q", v.OS, v.Channel.Track)
-}
